I/O cleanup + actually correct stuff
Copying a file path in Windows puts the path in quotes, which are annoying to remove manually each time. Now those are automatically removed before the file is accessed, so you don't have to manually take them out.

Moved a couple of lines around, in preparation for alpha sweeps. This should keep the code from asking for the same data for each angle of attack in that sweep.

Most importantly: VALIDATED SELIG DATA. Got root chord measurement of 0.400 m (feels reasonable) and used half-span of 0.724 m with an area of 0.233 m2 from the presented data, with an assumed rectangular wing. The results are remarkably similar, which will be posted soon.